+++ packages/cli/src/commands/commonOptions.ts
@@ -2,4 +2,5 @@ export interface CommonOptions {
   config?: string
   skipCodegen?: boolean
   whitelist?: string
+  skipWhitelist?: boolean
 }