+++ packages/cli/CHANGELOG.md
@@ -2,6 +2,11 @@
 
 ## Unreleased
 
+### Added
+
+- `options.whitelist` in PapiConfig to globally set the whitelist path.
+- `--skip-whitelist` CLI option to skip using a whitelist.
+
 ## 0.16.1 to 0.16.3 - 2025-11-27
 
 ### Fixed